+++ src/Symfony/Bundle/FrameworkBundle/Console/Descriptor/TextDescriptor.php
@@ -432,6 +432,7 @@ class TextDescriptor extends Descriptor
                         ['<info>Default value</>', $env['default_available'] ? $dump($env['default_value']) : 'n/a'],
                         ['<info>Real value</>', $env['runtime_available'] ? $dump($env['runtime_value']) : 'n/a'],
                         ['<info>Processed value</>', $env['default_available'] || $env['runtime_available'] ? $dump($env['processed_value']) : 'n/a'],
+                        ['<info>Usage count</>', $dump($env['usage_count'])],
                     ]);
                 }
             }
@@ -455,6 +456,7 @@ class TextDescriptor extends Descriptor
         $missing = [];
         foreach ($envs as $env) {
             if (isset($rows[$env['name']])) {
+                $rows[$env['name']][3] += $env['usage_count'];
                 continue;
             }
 
@@ -462,13 +464,20 @@ class TextDescriptor extends Descriptor
                 $env['name'],
                 $env['default_available'] ? $dump($env['default_value']) : 'n/a',
                 $env['runtime_available'] ? $dump($env['runtime_value']) : 'n/a',
+                $env['usage_count'],
             ];
             if (!$env['default_available'] && !$env['runtime_available']) {
                 $missing[$env['name']] = true;
             }
         }
 
-        $options['output']->table(['Name', 'Default value', 'Real value'], $rows);
+        $rows = array_map(function ($row) use ($dump) {
+            $row[3] = $dump($row[3]);
+
+            return $row;
+        }, $rows);
+
+        $options['output']->table(['Name', 'Default value', 'Real value', 'Usage count'], $rows);
         $options['output']->comment('Note real values might be different between web and CLI.');
 
         if ($missing) {