+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -17,6 +17,7 @@ CHANGELOG
  * Allow setting `debug.container.dump` to `false` to disable dumping the container to XML
  * Add `framework.http_cache.skip_response_headers` option
  * Display warmers duration on debug verbosity for `cache:clear` command
+ * Improved `bin/console debug:container --env-vars` and the `--env-var` option to display the number of occurrences of each environment variable in the container.
 
 6.2
 ---